changed testing framework back to mocha from jest
