+++ package.json
@@ -21,7 +21,8 @@
     "start-dev": "NODE_ENV='development' npm run build-client-watch & NODE_ENV='development' npm run start-server",
     "start-server": "nodemon server -e html,js,scss --ignore public --ignore client",
     "test": "NODE_ENV='test' mocha \"./server/**/*.spec.js\" \"./client/**/*.spec.js\" \"./script/**/*.spec.js\" --require @babel/polyfill --require @babel/register",
-    "jest": "jest --watch tests"
+    "mocha": "NODE_ENV='test' mocha \"./tests/*.spec.js\" --watch --require @babel/polyfill --require @babel/register",
+    "mocha:unit": "NODE_ENV='test' mocha --watch --require @babel/polyfill --require @babel/register"
   },
   "husky": {
     "hooks": {
@@ -80,7 +81,9 @@
     "@babel/preset-react": "^7.8.3",
     "@babel/register": "^7.8.3",
     "axios-mock-adapter": "^1.16.0",
+    "babel-core": "^6.26.3",
     "babel-eslint": "^8.2.6",
+    "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.5",
     "chai": "^4.2.0",
     "enzyme": "^3.11.0",
@@ -100,6 +103,7 @@
     "react-test-renderer": "^16.12.0",
     "redux-devtools-extension": "^2.13.5",
     "redux-mock-store": "^1.5.4",
+    "regenerator-runtime": "^0.13.3",
     "simple-git": "^1.129.0",
     "supertest": "^4.0.2",
     "webpack": "^4.41.5",
